geracion completa del modelo jpa

+++ src/com/nw/model/dao/EnvasadoControlPesoFillCabeceraDAO.java
@@ -6,7 +6,7 @@ import com.nw.model.EnvasadoControlPesoFillCabecera;
 
 public interface EnvasadoControlPesoFillCabeceraDAO {
 	public EnvasadoControlPesoFillCabecera updateEnvasadoControlPesoFillCabecera(EnvasadoControlPesoFillCabecera envasadoControlPesoFillCabecera);
-	public EnvasadoControlPesoFillCabecera getByProduccionTurnoOrden(Long idenvasadoproceso, Long idproducciondetalleorden);
+	public EnvasadoControlPesoFillCabecera getByProduccionTurnoOrden(Long idenvasadoproceso, Integer idproducciondetalleorden);
 	public List<EnvasadoControlPesoFillCabecera> getByProduccion(Long idenvasadoproceso);
 	public boolean eliminaEnvasadoControlPesoFillCabecera(EnvasadoControlPesoFillCabecera ecpfc);
 }
